renamed aws_utils to utils, used env var for jwt auth
