+++ src/api/utils/make_db.py
@@ -1,6 +1,6 @@
 import duckdb
 import pandas as pd
-from aws_utils.s3_utils import create_s3_conn_from_creds, download_from_s3, upload_to_s3
+from utils.s3_utils import create_s3_conn_from_creds, download_from_s3, upload_to_s3
 from datetime import datetime
 import numpy as np
 import os
